requirements.txt
django==2.2.10
python-gettext
ordereddict
psycopg2
django-localflavor==3.0
lxml
flup
paramiko
Pillow
django-extensions
django-registration
django-crum
python-stdnum
django-iban
pytz
gunicorn
django-sendgrid-v5
python-dotenv